requirements.txt
discord.py==2.4.0
yt_dlp==2024.8.6
